putmapping

+++ src/main/java/com/EMS/entity/Employee.java
@@ -7,20 +7,26 @@ import jakarta.persistence.Id;
 
 @Entity
 public class Employee {
+
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private String surname;
     private int wage;
 
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
-    public void setId(int id){
-        this.id=id;
+
+    public void setId(Long id) {
+        this.id = id;
     }
+
+
     public String getName(){
         return name;
     }
